+++ habapp_rules/actors/state_observer.py
@@ -3,6 +3,7 @@ from __future__ import annotations
 
 import abc
 import logging
+import threading
 import time
 import typing
 
@@ -22,13 +23,16 @@ CallbackType = typing.Callable[[EventTypes], None]
 class _StateObserverBase(HABApp.Rule, abc.ABC):
 	"""Base class for observer classes."""
 
-	def __init__(self, item_name: str, control_names: list[str] | None = None, group_names: list[str] | None = None):
+	def __init__(self, item_name: str, control_names: list[str] | None = None, group_names: list[str] | None = None, value_tolerance: int = 0):
 		"""Init state observer for switch item.
 
 		:param item_name: Name of observed item
 		:param control_names: list of control items.
 		:param group_names: list of group items where the item is a part of. Group item type must match with type of item_name
+		:param value_tolerance: used by all observers which handle numbers. It can be used to allow a difference when comparing new and old values.
 		"""
+		self._value_tolerance = value_tolerance
+
 		HABApp.Rule.__init__(self)
 		self._instance_logger = habapp_rules.core.logger.InstanceLogger(LOGGER, item_name)
 
@@ -123,6 +127,15 @@ class _StateObserverBase(HABApp.Rule, abc.ABC):
 		callback: CallbackType = getattr(self, cb_name)
 		callback(event)
 
+	def _compare_values_with_tolerance(self, value_1: float, value_2: float) -> bool:
+		"""Compare values with tolerance
+
+		:param value_1: first value
+		:param value_2: second value
+		:return: true if values are the same (including the offset), false if not
+		"""
+		return abs((value_1 or 0) - (value_2 or 0)) > self._value_tolerance
+
 
 class StateObserverSwitch(_StateObserverBase):
 	"""Class to observe the on/off state of a switch item.
@@ -151,7 +164,7 @@ class StateObserverSwitch(_StateObserverBase):
 		self._cb_on = cb_on
 		self._cb_off = cb_off
 		_StateObserverBase.__init__(self, item_name)
-		self._value = self._value
+		self._value = self._value # todo: this makes no sense!
 
 	def _check_manual(self, event: HABApp.openhab.events.ItemStateChangedEvent | HABApp.openhab.events.ItemCommandEvent) -> None:
 		"""Check if light was triggered by a manual action
@@ -218,7 +231,7 @@ class StateObserverDimmer(_StateObserverBase):
 			cb_brightness_change=callback_change)
 	"""
 
-	def __init__(self, item_name: str, cb_on: CallbackType, cb_off: CallbackType, cb_brightness_change: CallbackType | None = None, control_names: list[str] | None = None, group_names: list[str] | None = None) -> None:
+	def __init__(self, item_name: str, cb_on: CallbackType, cb_off: CallbackType, cb_brightness_change: CallbackType | None = None, control_names: list[str] | None = None, group_names: list[str] | None = None, value_tolerance: int = 0) -> None:
 		"""Init state observer for dimmer item.
 
 		:param item_name: Name of dimmer item
@@ -227,9 +240,10 @@ class StateObserverDimmer(_StateObserverBase):
 		:param cb_brightness_change: callback which is called if dimmer is on and brightness changed
 		:param control_names: list of control items. They are used to also respond to switch on/off via INCREASE/DECREASE
 		:param group_names: list of group items where the item is a part of. Group item type must match with type of item_name
+		:param value_tolerance: the tolerance can be used to allow a difference when comparing new and old values.
 		"""
 
-		_StateObserverBase.__init__(self, item_name, control_names, group_names)
+		_StateObserverBase.__init__(self, item_name, control_names, group_names, value_tolerance)
 
 		self._cb_on = cb_on
 		self._cb_off = cb_off
@@ -249,7 +263,7 @@ class StateObserverDimmer(_StateObserverBase):
 				self._value = 0
 				self._trigger_callback("_cb_off", event)
 
-			elif event.value != self._value:
+			elif self._compare_values_with_tolerance(event.value, self._value):
 				self._value = event.value
 				self._trigger_callback("_cb_brightness_change", event)
 
@@ -317,20 +331,22 @@ class StateObserverRollerShutter(_StateObserverBase):
 				)
 		"""
 
-	def __init__(self, item_name: str, cb_manual: CallbackType, control_names: list[str] | None = None, group_names: list[str] | None = None) -> None:
+	def __init__(self, item_name: str, cb_manual: CallbackType, control_names: list[str] | None = None, group_names: list[str] | None = None, value_tolerance: int = 0) -> None:
 		"""Init state observer for dimmer item.
 
 		:param item_name: Name of dimmer item
 		:param cb_manual: callback which is called if a manual interaction was detected
 		:param control_names: list of control items. They are used to also respond to switch on/off via INCREASE/DECREASE
 		:param group_names: list of group items where the item is a part of. Group item type must match with type of item_name
+		:param value_tolerance: the tolerance can be used to allow a difference when comparing new and old values.
 		"""
-		_StateObserverBase.__init__(self, item_name, control_names, group_names)
+		self._value_tolerance = value_tolerance
+		_StateObserverBase.__init__(self, item_name, control_names, group_names, value_tolerance)
 
 		self._cb_manual = cb_manual
 
 	def _check_manual(self, event: HABApp.openhab.events.ItemStateChangedEvent | HABApp.openhab.events.ItemCommandEvent) -> None:
-		if isinstance(event.value, (int, float)) and event.value != self._value:
+		if isinstance(event.value, (int, float)) and self._compare_values_with_tolerance(event.value, self._value):
 			self._value = event.value
 			self._trigger_callback("_cb_manual", event)
 
@@ -368,16 +384,15 @@ class StateObserverNumber(_StateObserverBase):
 	habapp_rules.actors.state_observer.StateObserverNumber("I01_01_Number", callback_value_changed)
 	"""
 
-	def __init__(self, item_name: str, cb_manual: CallbackType, threshold: float = 0.1):
+	def __init__(self, item_name: str, cb_manual: CallbackType, value_tolerance: int = 0) -> None:
 		"""Init state observer for switch item.
 
 		:param item_name: Name of switch item
 		:param cb_manual: callback which should be called if manual change was detected
-		:param threshold: difference threshold which will trigger cb_manual (Can be used to ignore errors from rounding)
+		:param value_tolerance: the tolerance can be used to allow a difference when comparing new and old values.
 		"""
 		self._cb_manual = cb_manual
-		self.__threshold = threshold
-		_StateObserverBase.__init__(self, item_name)
+		_StateObserverBase.__init__(self, item_name, value_tolerance=value_tolerance)
 
 	def _check_manual(self, event: HABApp.openhab.events.ItemStateChangedEvent | HABApp.openhab.events.ItemCommandEvent) -> None:
 		"""Check if light was triggered by a manual action
@@ -389,7 +404,7 @@ class StateObserverNumber(_StateObserverBase):
 			self._value = event.value
 			return
 
-		if abs(event.value - self._value) > self.__threshold:
+		if self._compare_values_with_tolerance(event.value, self._value):
 			self._value = event.value
 			self._trigger_callback("_cb_manual", event)
 
@@ -409,3 +424,47 @@ class StateObserverNumber(_StateObserverBase):
 			raise ValueError(f"The given value is not supported for StateObserverNumber: {value}")
 		self._value = value
 		self._item.oh_send_command(value)
+
+
+class StateObserverSlat(StateObserverNumber):
+	"""This is only used for the slat value of shading!"""
+
+	def __init__(self, item_name: str, cb_manual: CallbackType, value_tolerance: int = 0) -> None:
+		"""Init state observer for switch item.
+
+		:param item_name: Name of switch item
+		:param cb_manual: callback which should be called if manual change was detected
+		:param value_tolerance: the tolerance can be used to allow a difference when comparing new and old values.
+		"""
+		self.__timer_manual: threading.Timer | None = None
+		StateObserverNumber.__init__(self, item_name, cb_manual, value_tolerance)
+
+	def _check_manual(self, event: HABApp.openhab.events.ItemStateChangedEvent | HABApp.openhab.events.ItemCommandEvent) -> None:
+		"""Check if light was triggered by a manual action
+
+		:param event: event which triggered this method. This will be forwarded to the callback
+		:raises ValueError: if event is not supported
+		"""
+		self._stop_timer_manual()
+		if event.value in {0, 100}:
+			self.__timer_manual = threading.Timer(3, self.__cb_check_manual_delayed, [event])
+			self.__timer_manual.start()
+		else:
+			StateObserverNumber._check_manual(self, event)
+
+	def __cb_check_manual_delayed(self, event: HABApp.openhab.events.ItemStateChangedEvent | HABApp.openhab.events.ItemCommandEvent) -> None:
+		"""Trigger delayed manual check
+
+		:param event: event which should be checked
+		"""
+		StateObserverNumber._check_manual(self, event)
+
+	def _stop_timer_manual(self) -> None:
+		"""Stop timer if running."""
+		if self.__timer_manual:
+			self.__timer_manual.cancel()
+			self.__timer_manual = None
+
+	def on_rule_removed(self) -> None:
+		"""Stop timer if rule is removed."""
+		self._stop_timer_manual()