+++ changelog.md
@@ -11,6 +11,19 @@
 
 - fixed bug in ``habapp_rules.actors.state_observer.StateObserverNumber`` which triggered the manual-detected-callback if the received number deviates only a little bit because of data types. (e.g.: 1.000001 != 1.0) 
 
+# Version 5.4.3 - 14.01.2024
+
+## Bugfix
+
+- fixed bug in ``habapp_rules.actors.shading.Raffstore`` which triggered a hand detection also if only small slat differences occurred
+
+# Version 5.4.2 - 14.01.2024
+
+## Bugfix
+
+- fixed bug in all observers of ``habapp_rules.actors.state_observer`` which triggered the manual callback also if the value change of numeric values is tiny
+- fixed bug in ``habapp_rules.actors.shading._ShadingBase`` which triggered a hand detection also if only small position differences occurred
+
 # Version 5.4.1 - 26.12.2023
 
 ## Bugfix